feat(bet4-buildvis): extract build observability + evidence tools into scoped ToolPacks (EP-8DC217EB BET-4)
Process-Spine-Decision: BET-4 drain-to-zero (plan §4) — behavior-identical; imports shared build helpers.
BI: BI-2B7EE073
