+++ scripts/mcp-tool-pack-baseline.json
@@ -11,15 +11,10 @@
   "edit_sandbox_file",
   "execute_promotion",
   "generate_code",
-  "get_build_dispatch_history",
   "get_build_engine_readiness",
-  "get_build_progress_visibility",
   "get_build_sandbox_state",
-  "get_build_scoped_verification",
   "get_release_status",
-  "inspect_build_code_impact",
   "iterate_sandbox",
-  "list_build_activity_since",
   "list_sandbox_files",
   "principle_decide",
   "promote_to_build_studio",
@@ -30,7 +25,6 @@
   "read_sandbox_file",
   "reconcile_build_engines",
   "record_execution_evidence",
-  "record_external_development_evidence",
   "record_functional_failure_evidence",
   "record_local_integration_result",
   "recover_sandbox",
@@ -44,8 +38,6 @@
   "run_tool_script",
   "run_ux_test",
   "saveBuildEvidence",
-  "save_build_notes",
-  "save_phase_handoff",
   "schedule_promotion",
   "schedule_release_bundle",
   "search_sandbox",
